Simplify the worm's target direction logic
The worm's target direction was stored as a Maybe Direction, instead
of just a Direction. This was to signify the possibility of not
having a target direction, as a result of the user not pressing any
arrow key, in which case the worm should move in the current
direction. This, however, obscured the logic and didn't provide
anything over using just a Direction. Also, the field was renamed to
"targetDirection" to better express it's meaning.